'''CORE:''' Removed Signets/ Symbols Temporary.

+++ L2JTeon/java/net/sf/l2j/gameserver/skills/DocumentBase.java
@@ -385,8 +385,7 @@ abstract class DocumentBase
     {
 	Condition cond = null;
 	int[] ElementSeeds = new int[5];
-	// int[] forces = new int[2];
-        byte[] forces = new byte[2];
+	int[] forces = new int[2];
 	NamedNodeMap attrs = n.getAttributes();
 	for (int i = 0; i < attrs.getLength(); i++)
 	{
@@ -452,12 +451,10 @@ abstract class DocumentBase
 		ElementSeeds[4] = Integer.decode(getValue(a.getNodeValue(), null));
 	    } else if ("battle_force".equalsIgnoreCase(a.getNodeName()))
 	    {
-		// forces[0] = Integer.decode(getValue(a.getNodeValue(), null));
-                forces[0] = Byte.decode(getValue(a.getNodeValue(), null));
+		forces[0] = Integer.decode(getValue(a.getNodeValue(), null));
 	    } else if ("spell_force".equalsIgnoreCase(a.getNodeName()))
 	    {
-		// forces[1] = Integer.decode(getValue(a.getNodeValue(), null));
-                forces[1] = Byte.decode(getValue(a.getNodeValue(), null));
+		forces[1] = Integer.decode(getValue(a.getNodeValue(), null));
 	    }
 	}
 	// Elemental seed condition processing
